MUI styles comprised in one file
